+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+* **2013-08-16**: [Model] moved Image document, interface and logic to CmfMedia
+  , the `image.static_basepath` configuration is renamed to `image.basepath`
+
 1.0.0-beta3
 -----------